requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
requests==2.31.0
openai==1.58.0
pinecone-client==3.2.2
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
ItsDangerous==2.1.2
click==8.1.7
blinker==1.7.0
PyJWT==2.8.0
python-dotenv==1.0.0
